Your branch is up to date with 'origin/master'.

+++ init.sh
@@ -12,9 +12,14 @@ for file in $DIR/*; do
     filename=$(basename $file)
     if [[ "$filename" != "$(basename $0)" ]]; then
         if [ -e $HOME/.$filename ]; then
-            # remove existing dotfile
-            status "deleted existing file $HOME/.$filename"
-            rm $HOME/.$filename
+            if ! [[ -L $HOME/.$filename ]]; then
+              # remove existing dotfile
+              rm $HOME/.$filename
+              status "deleted existing file $HOME/.$filename"
+            else
+              status "symlink for .$filename exists"
+              continue
+            fi
         fi
         status "installing $filename --> .$filename"
         ln -s $DIR/$filename $HOME/.$filename
